Reset hardware related state on hardware disconnect
